fix: import resolver for web
