--- apps/desktop/src/renderer/src/components/ui/modal/stacked/modal-stack.desktop.tsx
@@ -1,31 +1,0 @@
-import { AnimatePresence } from "framer-motion"
-import { useAtomValue } from "jotai"
-
-import { modalStackAtom } from "./atom"
-import { useModalStack } from "./hooks"
-import { ModalInternal } from "./modal"
-import { useModalStackCalculationAndEffect } from "./modal-stack.shared"
-
-export const ModalStack = () => {
-  const { present } = useModalStack()
-  window.presentModal = present
-
-  const stack = useAtomValue(modalStackAtom)
-
-  const { topModalIndex, overlayOptions } = useModalStackCalculationAndEffect()
-
-  return (
-    <AnimatePresence mode="popLayout">
-      {stack.map((item, index) => (
-        <ModalInternal
-          key={item.id}
-          item={item}
-          index={index}
-          isTop={index === topModalIndex}
-          isBottom={index === 0}
-          overlayOptions={overlayOptions}
-        />
-      ))}
-    </AnimatePresence>
-  )
-}
