+++ apps/desktop/src/renderer/src/components/ui/modal/stacked/modal-stack.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { ModalStack as Desktop } from "./modal-stack.desktop"
+import { ModalStack as Desktop } from "./modal-stack.electron"
 import { ModalStack as Mobile } from "./modal-stack.mobile"
 
 export const ModalStack = withResponsiveSyncComponent(Desktop, Mobile)
