+++ apps/desktop/src/renderer/src/modules/app-layout/entry-content/mobile.tsx
@@ -1,6 +1,6 @@
 import { ROUTE_ENTRY_PENDING } from "~/constants"
 import { useRouteParamsSelector } from "~/hooks/biz/useRouteParams"
-import { EntryContent } from "~/modules/entry-content"
+import { EntryContent } from "~/modules/entry-content/index.mobile"
 
 import { EntryColumnMobile } from "../entry-column/mobile"
 
@@ -12,5 +12,6 @@ export const EntryContentMobile = () => {
   if (entryId === ROUTE_ENTRY_PENDING) {
     return <EntryColumnMobile />
   }
+
   return <EntryContent entryId={entryId!} />
 }
