+++ apps/desktop/src/renderer/src/modules/app-layout/subview/index.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { SubviewLayout as SubviewLayoutDesktop } from "./index.desktop"
+import { SubviewLayout as SubviewLayoutDesktop } from "./index.electron"
 import { SubviewLayout as SubviewLayoutMobile } from "./index.mobile"
 
 export const SubviewLayout = withResponsiveSyncComponent(SubviewLayoutDesktop, SubviewLayoutMobile)
