+++ apps/desktop/src/renderer/src/modules/discover/inbox-table.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { InboxTable as InboxTableDesktop } from "./inbox-table.desktop"
+import { InboxTable as InboxTableDesktop } from "./inbox-table.electron"
 import { InboxTable as InboxTableMobile } from "./inbox-table.mobile"
 
 export const InboxTable = withResponsiveSyncComponent<ComponentType>(
