+++ apps/desktop/src/renderer/src/modules/user/ProfileButton.tsx
@@ -1,7 +1,7 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { ProfileButton as ProfileButtonDesktop } from "./ProfileButton.desktop"
 import type { ProfileButtonProps } from "./ProfileButton.electron"
+import { ProfileButton as ProfileButtonDesktop } from "./ProfileButton.electron"
 import { ProfileButton as ProfileButtonMobile } from "./ProfileButton.mobile"
 
 export const ProfileButton = withResponsiveSyncComponent<ProfileButtonProps>(
