+++ apps/mobile/metro.config.js
@@ -22,7 +22,7 @@ config.resolver.resolveRequest = (context, moduleName, platform) => {
   const result = context.resolveRequest(context, moduleName, platform)
   if (result.type === "sourceFile") {
     const lastDotIndex = result.filePath.lastIndexOf(".")
-    const mobilePath = `${result.filePath.slice(0, lastDotIndex)}.mobile${result.filePath.slice(lastDotIndex)}`
+    const mobilePath = `${result.filePath.slice(0, lastDotIndex)}.rn${result.filePath.slice(lastDotIndex)}`
     const file = context.fileSystemLookup(mobilePath)
     if (file.exists) {
       return {
