+++ packages/shared/src/env.desktop.ts
@@ -1,3 +1,6 @@
+/**
+ * This env for apps/desktop
+ */
 import { createEnv } from "@t3-oss/env-core"
 import { z } from "zod"
